refactor: refactor SubscribeNewsletterForm and move it to components/ui
